astral: add type aliases to blueprints, blueprint sync, astral-hello round trip with greeting

+++ astral/struct_value_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 type testStruct struct {
-	SomeNumber int
+	SomeNumber int64
 	SomeList   []string
 }
